Incorporate DrByte's Optimization and further it.
Basically removed duplicate files from the admin making
the overall installation and upkeep easier after having
incorporated DrByte's optimization.  Code has been updated
to the extent known to support PHP 7.2 as well as maintain
compatibility with PHP 5.2.x for the time being.
It is expected that at some point such compatibility
will not be possible and will not be maintained.