Streamline bgcolor designation
In keeping with the concept of arrays for data handling
and all that is involved with the display format, added
an "html" type array to be able to possibly further modify
the look by making a single change instead of multiple
throughout the code. This was started with the "white"
background color. More may be done in the future to
flush out improved control.