Remove session information from the URL
This removes the session variable information from the
browser URL while maintaining the same level of "support".

If this does not function for the associated site,
recommendation is to verify that the page passes HTML
validation.  In the future this javascript code may get
moved out to a separate file or something; however, for
now is incorporated directly into the page.