Move database removal code to external code (observer)
Request had been made in the forum:
https://www.zen-cart.com/showthread.php?35081-User-tracking-mod&p=1363766#post1363766

To support the ability to "routinely" remove old data.
To make that possible, the delete code needed to be outside
of the main execution code, but still be executable from
within that code.

Without creating a new class the code was incorporated into
the observer.  This may change again in the future, but
offers the ability to call the removal code and remove
records based on the database settings...