Rework installer
Revised the $current_veresion variable to be specific
to the plugin (conflict exists/existed with SiteMap XML)

Cast some values to integers based on expected data type.

In implementation of 1.5.10, discovered that sorting logic
was not quite right. This is corrected here and will be
carried over to similar installers.

Refactored some of the code to provide early response and
easier code review/reading.

Finally, corrected logic for preventing version checking
when so desired.